consolidate getWorkspaceFolderFromTestUri
